pyproject.toml
[tool.poetry]
name = "nord-security-task"
version = "0.1.0"
description = "Nord Security home task"
authors = ["Farid Mirzayev <faridmirzayev@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.8.2"
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
dependency-injector = "^4.41.0"
boto3 = "^1.34.140"
botocore = "^1.34.140"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.1.0"
isort = "^5.13.2"
mypy = "^1.10.1"
pylint = "^3.2.5"
pre-commit = "^3.7.1"
ruff = "^0.5.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "C",  # mccabe rules
    "F",  # pyflakes rules
    "E",  # pycodestyle error rules
    "W",  # pycodestyle warning rules
    "B",  # flake8-bugbear rules
    "I",  # isort rules
]
ignore = [
    "C901",  # max-complexity-10
    "E501",  # line-too-long
]

[tool.ruff.format]
indent-style = "space"
quote-style = "single"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
